+++ packages/illustrations/src/__generated__/spotIcon/data/names.ts
@@ -1,6 +1,6 @@
 /**
  * DO NOT MODIFY
- * Generated from yarn nx run illustrations:sync
+ * Generated from yarn nx run illustration-tasks:sync-illustrations
  */
 
 import type { SpotIconName } from '../types/SpotIconName';