+++ packages/illustrations/src/__generated__/spotIcon/types/SpotIconName.ts
@@ -1,6 +1,6 @@
 /**
  * DO NOT MODIFY
- * Generated from yarn nx run illustrations:sync
+ * Generated from yarn nx run illustration-tasks:sync-illustrations
  */
 
 export type SpotIconName =